dom/workers/test/test_resolveWorker-assignment.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE html>
<html>
  <head>
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <script type="text/javascript">
      Worker = 17;

      var desc = Object.getOwnPropertyDescriptor(window, "Worker");
      is(typeof desc === "object" && desc !== null, "Worker property must exist");

      is(desc.value, 17, "Overwrite didn't work correctly");
      is(desc.enumerable, false,
         "Initial descriptor was non-enumerable, and [[Put]] changes the " +
         "property value but not its enumerability");
      is(desc.configurable, true,
         "Initial descriptor was configurable, and [[Put]] changes the " +
         "property value but not its configurability");
      is(desc.writable, true,
         "Initial descriptor was writable, and [[Put]] changes the " +
         "property value but not its writability");
    </script>
  </body>
</html>
